// frontend/src/shared/api/generated/model/chatMessageRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * novalearn-api
 * OpenAPI spec version: 0.1.0
 */
import type { FileAttachment } from './fileAttachment';
import type { InlineQuizAnswerContext } from './inlineQuizAnswerContext';

export interface ChatMessageRequest {
  conversation_id?: string | null;
  folder_id?: string | null;
  test_session_id?: string | null;
  question_id?: string | null;
  lesson_id?: string | null;
  current_block_id?: string | null;
  scope_type?: string | null;
  feedback_note_id?: string | null;
  inline_quiz_answers?: InlineQuizAnswerContext[] | null;
  /**
   * @minLength 1
   * @maxLength 150000
   */
  message: string;
  current_document_id?: string | null;
  current_page?: number | null;
  total_pages?: number | null;
  model?: string | null;
  reasoning?: string | null;
  images?: string[];
  citations?: string[];
  attachments?: FileAttachment[];
}